Scrub BaseTools C Source to solve some gaps between EFI BaseTypes to POSIX types for printf() and other libc functions. The following rules are applied: INT8, INT16   => "%d" with no type cast required.                            e.g. printf ("%d", Int16Value); UINT8, UINT16 => "%u","%x" or "%X" with no type cast required.               e.g. printf ("%x", Uint16Value); INT32         => "%d" with type-cast (int)                                   e.g. printf ("%d", (int) Int32Value); UINT32        => "%u","%x" or "%X" with type-cast (unsigned)                 e.g. printf ("%u", (unsigned) Uint32Value) INT64         => "%lld" with type-cast (long long)                           e.g. printf ("%lld", (long long) Int64Value); UINT64        => "%llu", "%llx" or "llX" with type-cast (unsigned long long) e.g. printf ("%llX", (unsigned long long) Uint64Value);
git-svn-id: https://edk2-buildtools.svn.sourceforge.net/svnroot/edk2-buildtools/trunk/BaseTools@1675 c2973edb-eda0-4c78-bc6a-9341b269661f

+++ Source/C/Common/MyAlloc.c
@@ -121,12 +121,12 @@ MyCheck (
       "\nFile=%s, Line=%u, nSize=%u, Head=%xh, Tail=%xh\n",
       Final,
       File,
-      (UINT32)Line,
+      (unsigned)Line,
       Tmp->File,
-      (UINT32)Tmp->Line,
-      (UINT32)Tmp->Size,
-      *(UINT32 *) (Tmp->Buffer),
-      *(UINT32 *) (&Tmp->Buffer[Tmp->Size + sizeof (UINT32)])
+      (unsigned) Tmp->Line,
+      (unsigned) Tmp->Size,
+      (unsigned) *(UINT32 *) (Tmp->Buffer),
+      (unsigned) *(UINT32 *) (&Tmp->Buffer[Tmp->Size + sizeof (UINT32)])
       );
 
     exit (1);
@@ -148,10 +148,10 @@ MyCheck (
         printf (
           "File=%s, Line=%u, nSize=%u, Head=%xh, Tail=%xh\n",
           Tmp->File,
-          (UINT32)Tmp->Line,
-          (UINT32)Tmp->Size,
-          *(UINT32 *) (Tmp->Buffer),
-          *(UINT32 *) (&Tmp->Buffer[Tmp->Size + sizeof (UINT32)])
+          (unsigned) Tmp->Line,
+          (unsigned) Tmp->Size,
+          (unsigned) *(UINT32 *) (Tmp->Buffer),
+          (unsigned) *(UINT32 *) (&Tmp->Buffer[Tmp->Size + sizeof (UINT32)])
           );
       }
     }
