+++ Source/C/GenBootSector/GenBootSector.c
@@ -262,9 +262,9 @@ Routine Description:
         } else {
           fprintf (
             stdout,
-            "%c: - DiskNum: %d, Type: %s\n", 
+            "%c: - DiskNum: %u, Type: %s\n", 
             DriveInfo.VolumeLetter,
-            DriveInfo.DiskNumber, 
+            (unsigned) DriveInfo.DiskNumber, 
             DriveInfo.DriveType->Description
             );
         }
@@ -732,7 +732,7 @@ main (
         return 1;
       }
       if (LogLevel > 9) {
-        Error (NULL, 0, 1003, "Invalid option value", "Debug Level range is 0-9, currnt input level is %d", LogLevel);
+        Error (NULL, 0, 1003, "Invalid option value", "Debug Level range is 0-9, currnt input level is %d", (int) LogLevel);
         return 1;
       }
       SetPrintLevel (LogLevel);
